YCP Killer: Unify YaST module structure
The existing directory tree layout of YaST modules wasn't consistent
across modules and it was different from the layout used after
installation. We decided to unify it to fit a regular structure. This
commit does so for this module.

The commit is a result of running YCP Killer's "restructure" command.

See also:

  https://github.com/yast/ycp-killer#new-yast-module-structure
  https://github.com/yast/ycp-killer